requirements.txt
# Dependências principais (versões compatíveis com Python 3.13)
opencv-contrib-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Face recognition (usando versões mais recentes)
# face-recognition>=1.3.0
# dlib>=19.22.0

# Utilidades
python-dotenv>=1.0.0

# Desenvolvimento e testes
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Dependências alternativas se face-recognition falhar
# Descomente se necessário:
# mediapipe>=0.10.0
# tensorflow>=2.13.0
